# PERF: MonthOffset.apply_index