- [x] closes #35048
- [ ] tests added / passed
- [x] passes `black pandas`
- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [ ] whatsnew entry

```
import pandas as pd
import numpy as np

N = 10000
rng = pd.date_range(start="1/1/2000", periods=N, freq="T")
offset = pd.offsets.MonthBegin()

In [6]: %timeit offset + rng
463 µs ± 19 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)   # <-- PR
508 µs ± 18.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <-- 1.0.4
736 µs ± 15 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <-- master
```